웨폰 아이템, 공격
<웨폰 아이템>
무기 등급 체계화, 교체 가능하게, 테이블 정리
<공격>
가변운 공격, 무거운 공격 입력 분리, 무기 별 애니메이션 따로, 등

+++ Assets/Scripts/Controllers/03.Monster/Monster.cs
@@ -69,7 +69,7 @@ public abstract  partial class Monster : AI
         m_rewards.Clear();
     }
 
-    public override void HitEvent(Character attacker, int dmg, bool isAnimation = true)
+    public override void HitEvent(Character attacker, float dmg, bool isAnimation = true)
     {
         base.HitEvent(attacker, dmg, isAnimation);
